backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-socketio
aiohttp
websockets
janus
requests
colorlog
pydantic-settings
